% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssgsea.R
\name{bccb_signature_names}
\alias{bccb_signature_names}
\title{Get a list of available signature names in the bccb_signatures set}
\usage{
bccb_signature_names()
}
\value{
Character vector with all expression signatures in the bccb_signatures set
}
\description{
Get a list of available signature names in the bccb_signatures set
}
