% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bccbRNASeq-package.R
\docType{package}
\name{bccbRNASeq-package}
\alias{bccbRNASeq}
\alias{bccbRNASeq-package}
\title{bccbRNASeq: Utility functions for RNASeq analysis from the Breast Oncology Center Computational Biology group at Dana Farber Cancer Institute}
\description{
Functions include: PAM50 assignments with cohort balancing, gene expression classification with cohort balancing, low gene expression filtering, ssGSEA, etc.
}
\keyword{internal}
