% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{classification_rescaled}
\alias{classification_rescaled}
\title{Get deciles/quantiles expression classification with cohort rebalancing}
\usage{
classification_rescaled(
  x,
  er_status,
  n_iter = 100,
  seed = 37,
  cores = getOption("mc.cores", 2L)
)
}
\arguments{
\item{x}{Numeric matrix with expression values}

\item{er_status}{Logical vector indicating ER positivity. NAs are treated as
Unknown or Indeterminate status}

\item{n_iter}{Number of iterations for sampling}

\item{seed}{Random seed}

\item{cores}{Number of cores for parallel processing}
}
\value{
Character matrix with the following categories: "Upper Decile",
"Upper Quartile", "Lower Decile", "Lower Quartile"
}
\description{
Get deciles/quantiles expression classification with cohort rebalancing
}
