% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterexpr.R
\name{expressed_genes}
\alias{expressed_genes}
\title{Get list of expressed genes plus genes of interest}
\usage{
expressed_genes(counts, min_count, min_total_count)
}
\arguments{
\item{counts}{Numeric matrix of gene counts. Rownames are gene symbols, columns are samples}

\item{min_count}{Minimum number of counts}

\item{min_total_count}{Minimum number of total counts}
}
\value{
A character vector of expressed genes.
}
\description{
Get list of expressed genes plus genes of interest
}
