% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterexpr.R
\name{filter_matrix}
\alias{filter_matrix}
\title{Filter counts and normalized counts (ex. tpms)  for expressed genes}
\usage{
filter_matrix(
  counts,
  min_count = 100,
  min_total_count = 150,
  genes_keep = NULL,
  norm_counts = NULL
)
}
\arguments{
\item{counts}{Numeric matrix of gene counts. Rownames are gene symbols, columns are samples}

\item{min_count}{Minimum number of counts}

\item{min_total_count}{Minimum number of total counts}

\item{genes_keep}{Character vector of genes to keep in the matrix regarding their counts}

\item{norm_counts}{Numeric matrix with normalized counts (ex. tpms). Should have the same dimensions as counts.}
}
\value{
List of size two with counts matrix and norm_counts filtered by expressed genes with genes in the BCCB curated list and
genes passed in the genes_keep parameter. If norm_counts are not provided, the second item is null
}
\description{
Filter counts and normalized counts (ex. tpms)  for expressed genes
}
