% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pam50.R
\name{genefu_pam50}
\alias{genefu_pam50}
\title{Get genefu predictions}
\usage{
genefu_pam50(x)
}
\arguments{
\item{x}{Numeric matrix with expression values}
}
\value{
Genefu object with PAM50 predictions
}
\description{
Get genefu predictions
}
