% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factors.R
\name{lufactors}
\alias{lufactors}
\title{Get lower and upper quantile and decile factors per rows}
\usage{
lufactors(x)
}
\arguments{
\item{x}{Numeric matrix of gene expression. Rows are genes, columns are samples}
}
\value{
A list of four vectors of factors per row:
\itemize{
\item{lq} {Lower quantile values}
\item{uq} {Upper quantile values}
\item{lq10} {Lower decile values}
\item{uq10} {Upper decile values}
}
}
\description{
Get lower and upper quantile and decile factors per rows
}
