% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterexpr.R
\name{missing_genes}
\alias{missing_genes}
\title{Get genes from our curated list missing from a counts matrix}
\usage{
missing_genes(counts)
}
\arguments{
\item{counts}{Numeric matrix of gene counts. Rownames are gene symbols, columns are samples}
}
\value{
Character vector of gene symbols
}
\description{
Get genes from our curated list missing from a counts matrix
}
