% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescale.R
\name{neg_missing}
\alias{neg_missing}
\title{Get number of ER negative samples to achieve a balanced cohort
according to J Clin Oncol. 2020 Dec 10; 38(35): 4184–4193.}
\usage{
neg_missing(er_status)
}
\arguments{
\item{er_status}{Logical vector indicating ER positivity. NAs are treated as
Unknown or Indeterminate status}
}
\value{
Number of ER negative additional samples required. If set has enough negative
samples, the function returns 0.
}
\description{
Get number of ER negative samples to achieve a balanced cohort
according to J Clin Oncol. 2020 Dec 10; 38(35): 4184–4193.
}
\examples{
neg_missing(sample(c(TRUE, FALSE, NA), size = 50, replace = TRUE, prob = c(0.9, 0.07, 0.03)))
}
