% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescale.R
\name{negative_samples_needed}
\alias{negative_samples_needed}
\title{Get number of negative samples to achieve a balanced cohort
From J Clin Oncol. 2020 Dec 10; 38(35): 4184–4193 for ER balance}
\usage{
negative_samples_needed(status, pos_ratio = 126/195)
}
\arguments{
\item{status}{Logical vector for status. NAs are treated as
Unknown or Indeterminate status}

\item{pos_ratio}{Numeric between 0 and 1 indicating the proportion of
positive samples wanted, without taking NAs into account}
}
\value{
Number of extra negative samples required. If negative, extra positive
samples are needed
}
\description{
Get number of negative samples to achieve a balanced cohort
From J Clin Oncol. 2020 Dec 10; 38(35): 4184–4193 for ER balance
}
\examples{
v  <- sample(c(TRUE, FALSE, NA), size = 50, replace = TRUE,
 prob = c(0.9, 0.07, 0.03))
negative_samples_needed(v, 0.8)
}
