% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pam50.R
\name{pam50_mixed}
\alias{pam50_mixed}
\title{Get PAM50 correlations}
\usage{
pam50_mixed(pam50_preds)
}
\arguments{
\item{pam50_preds}{Genefu object}
}
\value{
Data Frame with PAM50 correlations
}
\description{
Get PAM50 correlations
}
