% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pam50.R
\name{pam50_receptor_gene_centering}
\alias{pam50_receptor_gene_centering}
\title{Get pam50 predictions using gene centering.
This method is intended for single receptor cohorts.}
\usage{
pam50_receptor_gene_centering(x, receptor_status)
}
\arguments{
\item{x}{Numeric matrix with expression values}

\item{receptor_status}{Receptor status of the samples in the cohort. Accepted
values are: ERpos_HER2neg, HER2pos_ERneg, HER2pos_ERpos, TNBC}
}
\value{
Data frame with PAM50 predictions for each sample and its corresponding subtype correlations
}
\description{
Get pam50 predictions using gene centering.
This method is intended for single receptor cohorts.
}
