% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pam50.R
\name{pam50_rescaled}
\alias{pam50_rescaled}
\title{Get pam50 with rescalling}
\usage{
pam50_rescaled(
  x,
  er_status,
  n_iter = 1000,
  seed = 37,
  cores = getOption("mc.cores", 2L)
)
}
\arguments{
\item{x}{Numeric matrix with expression values}

\item{er_status}{Logical vector indicating ER positivity. NAs are treated as
Unknown or Indeterminate status}

\item{n_iter}{Number of iterations for sampling}

\item{seed}{Random seed}

\item{cores}{Number of cores for parallel processing}
}
\value{
Data frame with PAM50 predictions for each sample and its corresponding subtype correlations
}
\description{
Get pam50 with rescalling
}
