% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factors.R
\name{qfactors}
\alias{qfactors}
\title{Get factors for PAM50 normalization}
\usage{
qfactors(x)
}
\arguments{
\item{x}{Numeric matrix of gene expression. Rows are genes, columns are samples}
}
\value{
A list of two vectors of factors per row:
\itemize{
\item{maf} {ma factors}
\item{mif} {mi factors}
}
}
\description{
Get factors for PAM50 normalization
}
