% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pam50.R
\name{rescale_expr}
\alias{rescale_expr}
\title{Rescale expression matrix}
\usage{
rescale_expr(x, er_status, n_iter, seed, cores)
}
\arguments{
\item{x}{Numeric matrix with expression values}

\item{er_status}{Logical vector indicating ER positivity. NAs are treated as
Unknown or Indeterminate status}

\item{n_iter}{Number of iterations for sampling}

\item{seed}{Random seed}

\item{cores}{Number of cores for parallel processing}
}
\value{
Numeric rescaled as a balance ER cohort
}
\description{
Rescale expression matrix
}
