% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescale.R
\name{rescale_expr_median_center}
\alias{rescale_expr_median_center}
\title{Rescale expression matrix with user provided factors}
\usage{
rescale_expr_median_center(
  x,
  status,
  pos_ratio,
  qs,
  n_iter = 1000,
  seed = 37,
  cores = getOption("mc.cores", 2L)
)
}
\arguments{
\item{x}{Numeric matrix with expression values}

\item{status}{Logical vector for status. NAs are treated as
Unknown or Indeterminate status}

\item{pos_ratio}{Numeric between 0 and 1 indicating the proportion of
positive samples wanted, without taking NAs into account}

\item{qs}{Named vector for quantile probabilities.
Rownames in q and names in qs must match}

\item{n_iter}{Number of iterations for sampling}

\item{seed}{Random seed}

\item{cores}{Number of cores for parallel processing}
}
\value{
Numeric rescaled matrix as a balance cohort
}
\description{
Rescale expression matrix with user provided factors
}
