% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factors.R
\name{sfactors}
\alias{sfactors}
\title{Get factors}
\usage{
sfactors(x, qs)
}
\arguments{
\item{x}{Numeric matrix of gene expression. Rows are genes, columns are samples}

\item{qs}{Named vector for quantile probabilities.
Rownames in q and names in qs must match}
}
\value{
Named numeric vector with x row quantile factors
}
\description{
Get factors
}
