% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssgsea.R
\name{ss_gsea_par}
\alias{ss_gsea_par}
\title{Get single sample GSEA for a list of samples}
\usage{
ss_gsea_par(
  norm_counts,
  set_list,
  seed = 37,
  cores = getOption("mc.cores", 2L)
)
}
\arguments{
\item{norm_counts}{Numeric matrix with normalized counts. Rownames are gene symbols, columns are samples}

\item{set_list}{List of signatures for GSEA}

\item{seed}{Random seed}

\item{cores}{Number of cores for parallel processing}
}
\value{
Data frame with GSEA outputs from fgsea::fgsea. The data frame contains the following columns:
\itemize{
\item{"signature_e"}{}
\item{"NES"}{Normalized enrichment scores}
\item{"ES"}{Enrichment score}
\item{"pval"}{}
\item{"padj"}{}
\item{"size"}{}
\item{"leadingEdge"}{Genes in the leadingEdge vector}
\item{"sample"}{Sample id}
}
}
\description{
Get single sample GSEA for a list of samples
}
