% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pam50.R
\name{subset_pam50}
\alias{subset_pam50}
\title{Filter matrix to PAM50 genes}
\usage{
subset_pam50(x)
}
\arguments{
\item{x}{Numeric matrix with expression values}
}
\value{
Numeric matrix subsetted for PAM50 genes
}
\description{
Filter matrix to PAM50 genes
}
