% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterexpr.R
\name{expressed_genes}
\alias{expressed_genes}
\title{Get list of expressed genes plus genes of interest}
\usage{
expressed_genes(counts, min_count = 100, min_total_count = 150)
}
\arguments{
\item{counts}{Numeric matrix of gene counts. Rownames are gene symbols, columns are samples}

\item{min_count}{Minimum number of counts}

\item{min_total_count}{Minimum number of total counts}
}
\value{
A character vector of expressed genes.

This function uses the filterByExpr function from the edgeR Bioconductor package,
and re-adds genes of interest regarding of their expression values.
These include PAM50 genes, genes in commonly used
expression signatures and oncogenic genes.
}
\description{
Get list of expressed genes plus genes of interest
}
